iterm: Only install iterm prefs on MacOS

+++ setup.sh
@@ -6,6 +6,7 @@
 
 set -ex
 SOURCE_DIR=`pwd`
+OSNAME=$(uname -s)
 
 # Dot files
 cd ${SOURCE_DIR}/dot
@@ -44,10 +45,11 @@ for f in ${TOOLS_DIR}/*; do
 done
 
 # Misc
-
-# iterm preferences
-[ -f ~/Library/Preferences/com.googlecode.iterm2.plist ] && \
-    rm ~/Library/Preferences/com.googlecode.iterm2.plist
-ln -s ${SOURCE_DIR}/misc/com.googlecode.iterm2.plist ~/Library/Preferences/
-# Update them
-defaults read com.googlecode.iterm2
+if [ $OSNAME = "Darwin" ]; then
+    # iterm preferences
+    [ -f ~/Library/Preferences/com.googlecode.iterm2.plist ] && \
+        rm ~/Library/Preferences/com.googlecode.iterm2.plist
+    ln -s ${SOURCE_DIR}/misc/com.googlecode.iterm2.plist ~/Library/Preferences/
+    # Update them
+    defaults read com.googlecode.iterm2
+fi
